% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtime.R
\name{addtime}
\alias{addtime}
\title{Add time to BIOPAC AcqKnowledge files}
\usage{
addtime(x, time = 60)
}
\arguments{
\item{x}{An object containing data file from AcqKnowledge}

\item{time}{Specifies how long the entire file is in seconds. Default is 60 seconds.}
}
\value{
The original object is returned with a new field for time
}
\description{
Adds a field for time that has been dropped when AcqKnowledge exports a .TXT file
}
