% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findbpm.R
\name{findbpm}
\alias{findbpm}
\title{Determine heart rate}
\usage{
findbpm(x, t = 30)
}
\arguments{
\item{x}{A vector of voltages of an ecg from AcqKnowledge}

\item{t}{Specifies how long the entire file is in seconds. Default is 30 seconds.}
}
\value{
a single integer, the average heart rate in bpm, will be returned
}
\description{
Find heart rate from a ecg in BIOPAC AcqKnowledge file format.
}
