% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlybpm.R
\name{onlybpm}
\alias{onlybpm}
\title{Trims BIOPAC AcqKnowledge files to only include heart rate}
\usage{
onlybpm(x)
}
\arguments{
\item{x}{An object containing data file from AcqKnowledge which has already had time added by addtime function}
}
\value{
An object containining only time and heart rate at each heart beat is returned
}
\description{
AcqKnowledge exports all the voltage data from the electrodes when recording a heart rate, which is often more than what we want. This function is used to trim the data to only include heart rate.
}
