% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMP160.R
\name{read.mp160}
\alias{read.mp160}
\title{Read Data Files Exported from BioPac MP160}
\usage{
read.mp160(file)
}
\arguments{
\item{file}{File to be read}
}
\value{
Data will be read in, a column for time is added. Header information will be discarded.
}
\description{
Will read the output file from the MP160 logger software into R as a dataframe.
}
